llm-python/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.3
transformers==4.43.3
torch>=2.0.0
python-dotenv==1.0.1
pydantic==2.8.2
